Created testscript folder, moved constants and functions from root index to testscript folder.
