+++ src/js/testfiles/index.js
@@ -1,0 +1,2 @@
+export * from './constants.js';
+export * from './testScripts.js';
